requirements.txt
openpyxl==3.1.2
pyautogen==0.2.29
pydantic==2.7.4
PyPDF2==1.28.6